requirements.txt
# Core dependencies
pynput>=1.7.6          # For global input capture (minimum version for stability)
pillow>=10.0.0         # For screenshot and image processing
cryptography>=41.0.0   # For macro file encryption
pywin32>=306           # For Windows API access (GDI fallback)
# tkinter is built-in with Python

# Development dependencies
pytest>=7.0.0          # For testing
pytest-cov>=4.0.0     # For test coverage
ruff>=0.12.0           # For linting and formatting (replaces black and flake8)